+++ app/services/download_service.py
@@ -1,0 +1,428 @@
+"""Service layer for download-related database and worker interactions."""
+
+from __future__ import annotations
+
+from datetime import datetime
+from typing import Any, Dict, List, Optional
+
+from fastapi import HTTPException, status
+from sqlalchemy.orm import Session
+
+from app.core.transfers_api import TransfersApi, TransfersApiError
+from app.errors import AppError
+from app.logging import get_logger
+from app.models import Download
+from app.schemas import DownloadPriorityUpdate, SoulseekDownloadRequest
+from app.utils.activity import record_activity
+from app.utils.downloads import (
+    ACTIVE_STATES,
+    coerce_priority,
+    determine_priority,
+    render_downloads_csv,
+    resolve_status_filter,
+    serialise_download,
+)
+from app.utils.events import DOWNLOAD_BLOCKED
+from app.utils.service_health import collect_missing_credentials
+from app.workers.persistence import update_priority as update_worker_priority
+
+
+logger = get_logger(__name__)
+
+
+class DownloadService:
+    """Encapsulates persistence and worker coordination for downloads."""
+
+    def __init__(self, *, session: Session, transfers: TransfersApi) -> None:
+        self._session = session
+        self._transfers = transfers
+
+    # ------------------------------------------------------------------
+    # Query helpers
+    # ------------------------------------------------------------------
+    def _build_download_query(
+        self,
+        *,
+        include_all: bool,
+        status_filter: Optional[str] = None,
+        created_from: Optional[datetime] = None,
+        created_to: Optional[datetime] = None,
+    ) -> Any:
+        query = self._session.query(Download)
+        if status_filter:
+            states = resolve_status_filter(status_filter)
+            query = query.filter(Download.state.in_(tuple(states)))
+        elif not include_all:
+            query = query.filter(Download.state.in_(tuple(ACTIVE_STATES)))
+
+        if created_from:
+            query = query.filter(Download.created_at >= created_from)
+        if created_to:
+            query = query.filter(Download.created_at <= created_to)
+
+        return query.order_by(Download.priority.desc(), Download.created_at.desc())
+
+    def list_downloads(
+        self,
+        *,
+        include_all: bool,
+        status_filter: Optional[str],
+        limit: int,
+        offset: int,
+    ) -> List[Download]:
+        try:
+            query = self._build_download_query(
+                include_all=include_all, status_filter=status_filter
+            )
+            return query.offset(offset).limit(limit).all()
+        except (HTTPException, AppError):
+            raise
+        except Exception as exc:  # pragma: no cover - defensive database failure handling
+            logger.exception("Failed to list downloads: %s", exc)
+            raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail="Failed to fetch downloads",
+            ) from exc
+
+    def get_download(self, download_id: int) -> Download:
+        try:
+            download = self._session.get(Download, download_id)
+        except (HTTPException, AppError):
+            raise
+        except Exception as exc:  # pragma: no cover - defensive database failure handling
+            logger.exception("Failed to load download %s: %s", download_id, exc)
+            raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail="Failed to fetch download",
+            ) from exc
+
+        if download is None:
+            logger.warning("Download %s not found", download_id)
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Download not found")
+        return download
+
+    def update_priority(self, download_id: int, payload: DownloadPriorityUpdate) -> Download:
+        download = self.get_download(download_id)
+
+        new_priority = int(payload.priority)
+        download.priority = new_priority
+        download.updated_at = datetime.utcnow()
+        payload_copy = dict(download.request_payload or {})
+        payload_copy["priority"] = new_priority
+        download.request_payload = payload_copy
+
+        self._session.add(download)
+        self._session.commit()
+
+        logger.info("Updated priority for download %s to %s", download_id, new_priority)
+
+        job_id = download.job_id
+        if job_id and not update_worker_priority(job_id, new_priority, job_type="sync"):
+            logger.error(
+                "Failed to update worker job priority for download %s (job %s)",
+                download_id,
+                job_id,
+            )
+
+        return download
+
+    async def queue_downloads(
+        self,
+        payload: SoulseekDownloadRequest,
+        *,
+        worker: Any,
+    ) -> Dict[str, Any]:
+        if not payload.files:
+            logger.warning("Download request without files rejected")
+            raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="No files supplied")
+
+        missing_credentials = collect_missing_credentials(self._session, ("soulseek",))
+        if missing_credentials:
+            missing_payload = {service: list(values) for service, values in missing_credentials.items()}
+            logger.warning("Download blocked due to missing credentials: %s", missing_payload)
+            record_activity(
+                "download",
+                DOWNLOAD_BLOCKED,
+                details={"missing": missing_payload, "username": payload.username},
+            )
+            raise HTTPException(
+                status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
+                detail={"message": "Download blocked", "missing": missing_payload},
+            )
+
+        enqueue = getattr(worker, "enqueue", None) if worker else None
+        if enqueue is None:
+            logger.error("Download worker unavailable for request from %s", payload.username)
+            record_activity("download", "failed", details={"reason": "worker_unavailable"})
+            raise HTTPException(
+                status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
+                detail="Download worker unavailable",
+            )
+
+        download_records: List[Download] = []
+        job_files: List[Dict[str, Any]] = []
+        job_priorities: List[int] = []
+        try:
+            for file_info in payload.files:
+                filename = str(file_info.get("filename") or file_info.get("name") or "unknown")
+                priority = determine_priority(file_info)
+                download = Download(
+                    filename=filename,
+                    state="queued",
+                    progress=0.0,
+                    username=payload.username,
+                    priority=priority,
+                )
+                self._session.add(download)
+                self._session.flush()
+
+                payload_copy = dict(file_info)
+                payload_copy.setdefault("filename", filename)
+                payload_copy["download_id"] = download.id
+                payload_copy["priority"] = priority
+                download.request_payload = payload_copy
+                job_files.append(payload_copy)
+
+                download_records.append(download)
+                job_priorities.append(priority)
+            self._session.commit()
+        except Exception as exc:  # pragma: no cover - defensive persistence handling
+            self._session.rollback()
+            logger.exception("Failed to persist download request: %s", exc)
+            record_activity("download", "failed", details={"reason": "persistence_error"})
+            raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail="Failed to queue download",
+            ) from exc
+
+        job_priority = max(job_priorities or [0])
+        job = {"username": payload.username, "files": job_files, "priority": job_priority}
+        try:
+            await enqueue(job)  # type: ignore[func-returns-value]
+        except Exception as exc:  # pragma: no cover - defensive worker error
+            logger.exception("Failed to enqueue download job: %s", exc)
+            now = datetime.utcnow()
+            for download in download_records:
+                download.state = "failed"
+                download.updated_at = now
+            self._session.commit()
+            record_activity("download", "failed", details={"reason": "enqueue_error"})
+            raise HTTPException(
+                status_code=status.HTTP_502_BAD_GATEWAY,
+                detail="Failed to enqueue download",
+            ) from exc
+
+        record_activity(
+            "download",
+            "queued",
+            details={
+                "download_ids": [download.id for download in download_records],
+                "username": payload.username,
+            },
+        )
+
+        logger.info("Queued %d download(s) for %s", len(download_records), payload.username)
+
+        download_payload = [
+            {
+                "id": download.id,
+                "filename": download.filename,
+                "state": download.state,
+                "progress": download.progress,
+                "priority": download.priority,
+                "username": download.username,
+            }
+            for download in download_records
+        ]
+
+        response: Dict[str, Any] = {"status": "queued", "downloads": download_payload}
+        if download_records:
+            response["download_id"] = download_records[0].id
+        return response
+
+    async def cancel_download(self, download_id: int) -> Dict[str, Any]:
+        download = self.get_download(download_id)
+
+        if download.state not in {"queued", "running", "downloading"}:
+            logger.warning(
+                "Cancellation rejected for download %s due to invalid state %s",
+                download_id,
+                download.state,
+            )
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail="Download cannot be cancelled in its current state",
+            )
+
+        try:
+            await self._transfers.cancel_download(download_id)
+        except TransfersApiError as exc:
+            logger.error("slskd cancellation failed for %s: %s", download_id, exc)
+            raise HTTPException(
+                status_code=status.HTTP_502_BAD_GATEWAY,
+                detail="Failed to cancel download via slskd",
+            ) from exc
+
+        download.state = "cancelled"
+        download.updated_at = datetime.utcnow()
+
+        try:
+            self._session.commit()
+        except Exception as exc:  # pragma: no cover - defensive persistence handling
+            self._session.rollback()
+            logger.exception("Failed to persist cancellation for download %s: %s", download_id, exc)
+            raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail="Failed to cancel download",
+            ) from exc
+
+        record_activity(
+            "download",
+            "download_cancelled",
+            details={"download_id": download_id, "filename": download.filename},
+        )
+
+        return {"status": "cancelled", "download_id": download_id}
+
+    def export_downloads(
+        self,
+        *,
+        status_filter: Optional[str],
+        created_from: Optional[datetime],
+        created_to: Optional[datetime],
+        format: str,
+    ) -> ResponsePayload:
+        try:
+            query = self._build_download_query(
+                include_all=True,
+                status_filter=status_filter,
+                created_from=created_from,
+                created_to=created_to,
+            )
+            downloads = query.all()
+        except (HTTPException, AppError):
+            raise
+        except Exception as exc:  # pragma: no cover - defensive database failure handling
+            logger.exception("Failed to export downloads: %s", exc)
+            raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail="Failed to export downloads",
+            ) from exc
+
+        payload = [serialise_download(item) for item in downloads]
+        if format == "csv":
+            return {"content": render_downloads_csv(payload), "media_type": "text/csv"}
+        return {
+            "content": payload,
+            "media_type": "application/json",
+        }
+
+    async def retry_download(self, download_id: int) -> Dict[str, Any]:
+        original = self.get_download(download_id)
+
+        if original.state not in {"failed", "cancelled"}:
+            logger.warning(
+                "Retry rejected for download %s due to invalid state %s",
+                download_id,
+                original.state,
+            )
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail="Download cannot be retried in its current state",
+            )
+
+        if not original.username or not original.request_payload:
+            logger.error("Retry rejected for download %s due to missing payload", download_id)
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Download cannot be retried because original request data is missing",
+            )
+
+        payload_copy = dict(original.request_payload or {})
+        filename = payload_copy.get("filename") or original.filename
+        if not filename:
+            logger.error("Retry rejected for download %s due to missing filename", download_id)
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="Download cannot be retried because filename is unknown",
+            )
+
+        filesize = (
+            payload_copy.get("filesize")
+            or payload_copy.get("size")
+            or payload_copy.get("file_size")
+        )
+        if filesize is not None:
+            payload_copy.setdefault("filesize", filesize)
+
+        priority = coerce_priority(payload_copy.get("priority"))
+        if priority is None:
+            priority = original.priority
+
+        new_download = Download(
+            filename=filename,
+            state="queued",
+            progress=0.0,
+            username=original.username,
+            priority=priority,
+        )
+        self._session.add(new_download)
+        self._session.flush()
+
+        payload_copy["download_id"] = new_download.id
+        payload_copy.setdefault("filename", filename)
+        payload_copy["priority"] = priority
+        new_download.request_payload = payload_copy
+
+        try:
+            await self._transfers.cancel_download(download_id)
+        except TransfersApiError as exc:
+            self._session.rollback()
+            logger.error("slskd cancellation before retry failed for %s: %s", download_id, exc)
+            raise HTTPException(
+                status_code=status.HTTP_502_BAD_GATEWAY,
+                detail="Failed to cancel existing download via slskd",
+            ) from exc
+
+        try:
+            await self._transfers.enqueue(username=original.username, files=[payload_copy])
+        except TransfersApiError as exc:
+            self._session.rollback()
+            logger.error("Failed to enqueue retry for download %s: %s", download_id, exc)
+            raise HTTPException(
+                status_code=status.HTTP_502_BAD_GATEWAY,
+                detail="Failed to enqueue download via slskd",
+            ) from exc
+        except Exception as exc:  # pragma: no cover - defensive unexpected failure
+            self._session.rollback()
+            logger.exception("Unexpected error while retrying download %s: %s", download_id, exc)
+            raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail="Failed to retry download",
+            ) from exc
+
+        try:
+            self._session.commit()
+        except Exception as exc:  # pragma: no cover - defensive persistence handling
+            self._session.rollback()
+            logger.exception("Failed to persist retry download for %s: %s", download_id, exc)
+            raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail="Failed to retry download",
+            ) from exc
+
+        record_activity(
+            "download",
+            "download_retried",
+            details={
+                "original_download_id": download_id,
+                "retry_download_id": new_download.id,
+                "username": original.username,
+                "filename": filename,
+            },
+        )
+
+        return {"status": "queued", "download_id": new_download.id}
+
+
+ResponsePayload = Dict[str, Any]
